+++ api/server/server.go
@@ -2,13 +2,11 @@ package server
 
 import (
 	"bufio"
-	"bytes"
 	"context"
 	"crypto/tls"
 	"errors"
 	"fmt"
 	"io"
-	"net"
 	"net/http"
 	"os"
 	"path"
@@ -23,7 +21,6 @@ import (
 	"github.com/fnproject/fn/api/agent/hybrid"
 	"github.com/fnproject/fn/api/common"
 	"github.com/fnproject/fn/api/datastore"
-	"github.com/fnproject/fn/api/id"
 	"github.com/fnproject/fn/api/logs"
 	"github.com/fnproject/fn/api/models"
 	"github.com/fnproject/fn/api/mqs"
@@ -735,7 +732,6 @@ func New(ctx context.Context, opts ...Option) *Server {
 
 	}
 
-	setMachineID()
 	s.Router.Use(loggerWrap, traceWrap) // TODO should be opts
 	optionalCorsWrap(s.Router)          // TODO should be an opt
 	apiMetricsWrap(s)
@@ -965,39 +961,6 @@ func getPidList() ([]int, error) {
 		}
 	}
 	return pids, nil
-}
-
-func setMachineID() {
-	port := uint16(getEnvInt(EnvPort, DefaultPort))
-	addr := whoAmI().To4()
-	if addr == nil {
-		addr = net.ParseIP("127.0.0.1").To4()
-		logrus.Warn("could not find non-local ipv4 address to use, using '127.0.0.1' for ids, if this is a cluster beware of duplicate ids!")
-	}
-	id.SetMachineIdHost(addr, port)
-}
-
-// whoAmI searches for a non-local address on any network interface, returning
-// the first one it finds. it could be expanded to search eth0 or en0 only but
-// to date this has been unnecessary.
-func whoAmI() net.IP {
-	ints, _ := net.Interfaces()
-	for _, i := range ints {
-		if i.Name == "docker0" || i.Name == "lo" {
-			// not perfect
-			continue
-		}
-		addrs, _ := i.Addrs()
-		for _, a := range addrs {
-			ip, _, err := net.ParseCIDR(a.String())
-			if a.Network() == "ip+net" && err == nil && ip.To4() != nil {
-				if !bytes.Equal(ip, net.ParseIP("127.0.0.1")) {
-					return ip
-				}
-			}
-		}
-	}
-	return nil
 }
 
 func extractFields(c *gin.Context) logrus.Fields {
